confirm before delete reported item

+++ src/Pages/Dashboard/AdminDashboard/ReportedItems/ReportedItems.js
@@ -35,6 +35,14 @@ const ReportedItems = () => {
   }
 
   const handleDeleteProducts = (id) => {
+    const isConfirm = window.confirm(
+      "Are you sure you want to delete this product"
+    );
+
+    if (!isConfirm) {
+      return;
+    }
+
     setIsDataLoading(true);
     fetch(`https://thrift-store-server.vercel.app/reported-products/${id}`, {
       method: 'DELETE',
